# gh-129813: Use `PyBytesWriter` in `_json:_match_number_unicode`; `json.loads` with `int` or `float` 1.14x faster