The change is very similar to #138956

I'm surprised with the speed up:

# Benchmark

<details>
<summary>The script:</summary>

```python
import pyperf
import json

runner = pyperf.Runner()

num_elements = 1000

int_data = json.dumps(list(range(num_elements)))
runner.bench_func(f'json_loads_{num_elements}_ints', json.loads, int_data)

float_data = json.dumps([float(i) for i in range(num_elements)])
runner.bench_func(f'json_loads_{num_elements}_floats', json.loads, float_data)
```
</details>

The results (with `--rigorous`):

```
| Benchmark              | json-main | json-pybytes          |
|------------------------|:---------:|:---------------------:|
| json_loads_1000_ints   | 57.1 us   | 47.3 us: 1.21x faster |
| json_loads_1000_floats | 91.0 us   | 81.9 us: 1.11x faster |
| Geometric mean         | (ref)     | 1.16x faster          |
```

Where `main` is 4e00e2504fdcf06b5d55a773b31f49b1d3d6d35c

The environment:

```
% ./python -c "import sysconfig; print(sysconfig.get_config_var('CONFIG_ARGS'))"
'--enable-lto' '--with-optimizations'
```

`sudo ./python -m pyperf system tune` ensured.