@maurycy 2025-09-16
cc @vstinner 👋